### 2025-03-31 #7960 @editorialbot
[PRE REVIEW]: AutoUncertainties: A Python Package for Uncertainty Propagation
**Submitting author:** @cometbeetle (<a href="http://orcid.org/0009-0003-8206-5050">Ethan Mentzer</a>)
**Repository:** https://github.com/varchasgopalaswamy/AutoUncertainties
**Branch with paper.md** (empty if default branch):
**Version:** v0.6.3
