### 2025-04-02 #7960 @osorensen
:wave: @salrm8 @JakobBD @pescap @hanyyi would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
