### 2025-04-11 #7960 @osorensen
@damar-wicaksono @naingeet @dannys4 would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
