### 2025-04-16 varchasgopalaswamy/AutoUncertainties#21 @damar-wicaksono
JOSS Review (Reviewer #3): Software Paper
This issue is related to the [JOSS submission review](https://github.com/openjournals/joss-reviews/issues/8037).

Please find below some questions, remarks, and suggestions regarding the draft paper. Feel free to let me know if you need clarification. Thanks!

## Statement of need

- l. 13 "`AutoUncertainties` is **a** Python package uncertainty propagation.": Missing an article.

## Prior Work

- ll. 24-26 "While extremely useful, the `uncertainties` package relies on hand-implemented rules and functions for uncertainty propagation of array and scalar data."

In my opinion, it is not readily clear what those rules are and the problems associated with them. It will only become clear after the example below.

- ll. 26-27 "While this is transparent for the intrinsic dunder methods such as `__add__`, it becomes problematic for advanced mathematical operators."

I'm not sure I understand the term "transparent" here; does it mean "properly defined"? Furthermore, it would be more intuitive here to refer to these dunder methods as arithmetic (e.g., subtraction, multiplication, division) and logical operations. Finally, what are examples of "advanced mathematical operators" referred to here? The subsequent example mentions the use of cosine, which I think, is a function not an operator.

- If another package is used as an example in the paper, the relevant version should be mentioned (perhaps in the footnote).

- Not explicitly stated from the illustration here: One important limitation of using `uncertainties` with NumPy arrays, as illustrated in this example, is that while arrays of `ufloat` elements are supported and can be used with basic arithmetic operations, they cannot be directly passed to common NumPy functions. Instead, separate hand-implemented rules and functions provided by the `uncertainties` package must be used, i.e., there's no seamless integration.

## Implementation

- Could you please clarify and verify the formula, specifically with regard to the indices $i$ and $j$? Additionally, I believe the formula may be missing a summation over all dimensions ($i$). Furthermore, I suspect that this formula relies on an assumption of independence between variables, which I have mentioned elsewhere. Could you please confirm or correct these points?

- This section introduces at least two types of objects provided by `AutoUncertainties`: `Uncertainty` and `VectorUncertainty` (l. 44). However, other parts of the documentation also mention `ScalarUncertainty` and NumPy arrays with `ScalarUncertainty` elements. Could you clarify the relationships between these different objects and how they fit into the overall framework of `AutoUncertainties`?

- Building on my previous point, I notice that the example in the Prior Work section uses a `NumPy` array with `ufloat` elements from the `uncertainties` package, whereas the Implementation section uses instances of `ScalarUncertainty` and `VectorUncertainty`. Could you please clarify the similarities and differences between `VectorUncertainty` objects and `NumPy` arrays with `ScalarUncertainty` elements? Specifically, when should each be used, and are they interchangeable in certain contexts?

- One thing that would strengthen this section is an example illustrating how `AutoUncertainties` addresses the limitations of the `uncertainties` package, specifically the issue of using common NumPy functions like cosine. A concrete demonstration of how `AutoUncertainties` enables seamless usage of these functions, in contrast to the hand-implemented rules and functions required with `uncertainties`, would help to clearly showcase the benefits of the new package.

- I'm not entirely clear on the purpose of these accessors (`nominal_values()` and `std_devs()`). Could you please clarify their intended use case? It seems that the same information can be accessed directly through the `value` and `error` properties of an instance, so I'm wondering what additional benefits these accessors provide.

- It's not immediately apparent that 'error' and 'standard deviation' are being used interchangeably in this context. While it's later clarified that this equivalence assumes (independent) normally distributed variables, it might be helpful to introduce this assumption earlier in the paper. This would provide a clearer foundation for the subsequent discussion and avoid potential confusion.

- I think it's unnecessary to provide an example of how to change the rounding behavior in the display. Instead, it would be sufficient to simply state the default rounding rules (which, I assume, only affect the display, not the underlying values) and mention that users can customize this behavior if needed.

- ll. 56 - 59: Is this particular behavior of the package important enough to be in the paper? Perhaps such details in the documentation would be sufficient.

### Support for Pint

- Is the public class method `from_quantities()` necessary? I check the code and it seems the default constructor of `Uncertainty` already accepts `pint.Quantity` objects both for the value and the error.

### Pandas

- l. 68 "Upcoming **aditions** to `AutoUncertainties`...": Typo

## Current Limitations and Future Work

### Dependent Random Variables

- As mentioned before perhaps, the assumption on independence and normal distribution of the variables should be mentioned earlier.

- ll. 73-75 "This means that, in the case where the programmer **assumes dependence** between two or more Uncertainty objects, unexpected and counter-intuitive behavior may arise."

I'd like to raise an additional point related to the independence assumption. Besides the scenario where the programmer would like to assume dependence between `Uncertainty` objects (but they currently can't with `AutoUncertainties`), I think there's another potential issue when the programmer initially assumes independence but then inconsistently manipulates the variables in a way that implies dependence. The paper's example on subtraction touches on this, but I believe this problem generalizes to most, if not all, operations involving the same uncertain variables. For instance, adding the same uncertain variable (`X + X`) would violate the independence assumption if the quadrature formula is applied. In this case, the correct approach should be to treat it as `2X` (with the scaling rule), which `AutoUncertainties` is also able to handle.

It's worth emphasizing that this is a common pitfall. Although `AutoUncertainties` doesn't prevent users from performing operations like adding, subtracting, multiplying, or dividing identical variables, the resulting propagated uncertainty computed by the quadrature rule will be invalid due to the violated independence assumption. Pointing out this potential issue could help users avoid incorrect results.

- l. 77 "...will not result in a standard **devation** of zero...": Typo.

- ll. 80-83 "When multiplying a vector by a scalar Uncertainty object, each component of the resulting vector is assumed to be a multivariate normal distribution with no covariance, which may not be the desired behavior. For instance, taking the mean of such a vector will return an Uncertainty object with an unexpectedly small standard deviation."

I think this example should be clarified. In the code example:

- the result of `50 +/- 1.58114` is what expected if all elements of the vector are independent
- the result of `50 +/- 5` is what expected if all elements are fully correlated (e.g., each is basically the same variable)

However, I think it's worth noting that handling cases with covariance between independent and fully correlated variables would be more complex. In such scenarios, specifying the correlation between pairs of variables in an array would be necessary to accurately propagate the uncertainty. Perhaps it would be helpful to acknowledge this nuance.

- I would suggest to put the possible workaround to handle particular case of dependence in the documentation and simply mention that some possible workarounds can be found in the documentation.

### Typing System

Perhaps it won't be necessary to point out that full compliant with Mypy is the current limitation of `AutoUncertainties` in the paper.
