### 2025-04-16 varchasgopalaswamy/AutoUncertainties#22 @damar-wicaksono
JOSS Review (Reviewer #3): Functionality
This issue is related to the [JOSS submission review](https://github.com/openjournals/joss-reviews/issues/8037).

Please find below some questions, remarks, and suggestions regarding the draft paper. Feel free to let me know if you need clarification.

If there any other remarks or questions regarding the functionality of the package, I will put them here.

Thanks!

## Installation

- [x] I was having trouble installing the package following both README and the online documentation:

```
pip install auto_uncertainties
```

The command works, but it seems I get the wrong version of `AutoUncertainties`. In PyPi, the package is named [`auto-uncertainties`](https://pypi.org/project/auto-uncertainties/) and only after installing this I can follow the examples. Perhaps this can be clarified?

## Functionality

- [x] What should the relative uncertainty be when the central value is $0$?

For `ScalarUncertainty`, the result seems to be consistent with the definition:

```
>>> from auto_uncertainties import Uncertainty
>>> u = Uncertainty(0, 0.1)
>>> u.rel
nan
```

But for `VectorUncertainty`:

```
>>> uu = Uncertainty.from_sequence([u, u])
>>> uu.rel
array([0., 0.])
```

- [x] What is the default behavior for rounding in the display of an `Uncertainty` object?

In the documentation: "This can be turned off using `set_display_rounding`" which implies that the default is on. But:

```
>>> from auto_uncertainties import Uncertainty
>>> u = Uncertainty(4.514, 1.2483579)
>>> u
4.514 +/- 1.24836
>>> from auto_uncertainties import set_display_rounding
>>> set_display_rounding(False)
>>> u
4.514 +/- 1.24836  # The same as before
>>> set_display_rounding(True)
>>> u
4.5 +/- 1.2  # Rounded
```

Furthermore, it seems the default display behavior is already rounded following the default behavior of "g" format specifier that keeps 6 significant figures. Perhaps this should be clarified.

- [x] From the documentation "If [`numpy.array`](https://numpy.org/doc/stable/reference/generated/numpy.array.html#numpy.array "(in NumPy v2.2)") is called on an [`Uncertainty`](https://autouncertainties.readthedocs.io/en/latest/api/auto_uncertainties/uncertainty/uncertainty_containers/index.html#auto_uncertainties.uncertainty.uncertainty_containers.Uncertainty "auto_uncertainties.uncertainty.uncertainty_containers.Uncertainty") object, it will automatically get cast down to a [`numpy`](https://numpy.org/doc/stable/reference/index.html#module-numpy "(in NumPy v2.2)")array (and lose uncertainty information!), and emit a warning."

But there are two types of `Uncertainty` objects (If I am not wrong): `ScalarUncertainty` and `VectorUncertainty`, and they behave differently when passed into `np.array`:

```
>>> u = Uncertainty(1.0, 0.1)
>>> np.array(u)
array(1 +/- 0.1, dtype=object)   # does not get cast down
>>> uu = Uncertainty(np.array([1.0, 5.0]), np.array([0.1, 0.5]))
>>> np.array(uu)
<ipython-input-229-0dd2de1d7142>:1: DeprecationWarning: __array__ implementation doesn't accept a copy keyword, so passing copy=False failed. __array__ must implement 'dtype' and 'copy' keyword arguments. To learn more, see the migration guide https://numpy.org/devdocs/numpy_2_0_migration_guide.html#adapting-to-changes-in-the-copy-keyword
  np.array(uu)
<ipython-input-229-0dd2de1d7142>:1: DowncastWarning: The uncertainty is stripped when downcasting to ndarray.
  np.array(uu)
array([1., 5.]). # get cast down
```

Perhaps this edge case should be clarified.

- [x] I am curious about the practical differences between using `VectorUncertainty` and NumPy arrays with elements of `ScalarUncertainty`. Since both approaches involve elements of type `ScalarUncertainty`, what are the implications for their behavior in practice?

- [x] Possibly an edge case: The default constructor of `Uncertainty` accepts any type of `Sequence` for the value, but not for the error.

For instance, using a list for the value works as expected:

```
>>> vv = Uncertainty([2.0, 3.0])
>>> vv
[2 +/- 0, 3 +/- 0]  # assumes zero error
```

However, it seems counterintuitive that we can't also specify the errors as a sequence of the same length. For example:

```
>>> vv = Uncertainty([2.0, 3.0], [0.5, 0.6])
>>> vv
[2 +/- 0, 3 +/- 0]  # Does not raise exception, but ignore the specified errors
```

Wouldn't it be more intuitive to allow specifying two sequences for the values and errors to create a `VectorUncertainty`?

-  [x] Possibly an edge case: When an uncertain variable has a zero central value, its relative error is NaN (Not a Number). This can cause issues in multiplication and division operations, as the propagation formula relies on the relative error.

Consider the following:

```
>>> u = Uncertainty(0.0, 0.5)
>>> u.rel
nan
>>> v = Uncertainty(1.0, 0.1)
>>> v.rel
0.1
>>> u * v
0 +/- 0
```

The propagated error from the product is unexpectedly 0, even though the internal computation of the new error results in NaN. This is because the construction of an `Uncertainty` object with NaN as the error always sets the error to 0:

```
>>> Uncertainty(1.0, np.nan)
1 +/- 0
```

Upon investigation, I found that there is a check in place that sets any non-finite error (including NaN) to zero. In contrast, passing a negative error raises an exception.

If I am not wrong, even when the nominal value of a variable is zero, multiplication can still be computed using the general rule (i.e., the one involving derivatives) instead of the product rule, which relies on relative error.

I think it would be helpful to clarify this behavior.

## Performance

The submission does not make any performance claim.
