### 2025-04-18 varchasgopalaswamy/AutoUncertainties#22 @cometbeetle
Hi @damar-wicaksono— thanks for your comments! We appreciate your taking the time to review this.

The majority of things you mentioned make sense, and seem like good things to adjust in AutoUncertainties. As we work on some of the improvements, I will let you know if something isn't clear.

One thing I'll mention now is that the issue of instantiating `Uncertainty` with a sequence of values and errors (where the errors get ignored) is a bug. That, and some other potential bugs you described, will be fixed in [this PR](https://github.com/varchasgopalaswamy/AutoUncertainties/pull/26) at some point soon.
