### 2025-04-29 varchasgopalaswamy/AutoUncertainties#24 @JakobBD
JOSS Review (Reviewer #1)
This issue is related to the [JOSS submission review](https://github.com/openjournals/joss-reviews/issues/8037).

I have conducted this review independently from the other reviewers, so there may be redundancy in the comments.

### General impression

The paper is well written. The scope of the code meets JOSS guidelines.
The documentation is clear and comprehensive.
I see some issues, which are listed below.

### Main concern

The paper clearly and nicely states limitations of the package. However, these limitations are not found in the README and documentation of the package, which does not give potential users the chance to take them into consideration when deciding to use the package.

This applies especially to the fact that the uncertainties of two variables which are derived from the same uncertain variable are considered as independent. This limitation severely limits the applicability of the package as a black box add-on to existing code.

With my limited insight, I think the following might be a potential strategy to remedy this, albeit with high implementation effort:
  - for each uncertain variable, store a list of independent uncertainties
  - give each independent uncertainty a unique identifier at introduction
  - for operations between two variables, apply the operation to each uncertainty separately, where a correlation coefficient of one is assumed for uncertainties with the same identifier, and zero for differing ones
  - implement a function for calculation of the total error of a variable

If such a strategy is out of scope, then at least the limitation should be easily accessible to potential users, preferably by copying the relevant section of the paper into the README.

### Other minor issues

- The license file only contains the reference to a license, and could/should be replaced with a full license file.
- If the package is installed via pip, then `from auto_uncertainties import set_downcast_error` throws an error. However, `from auto_uncertainties import Uncertainty` and `from auto_uncertainties.uncertainty import set_downcast_error` both work. Seems to be a package building issue.
- The modulo operation quietly sets the standard deviation to zero. This is different from other non-linear operators, such as `abs`, which leaves the error unchanged. I suppose leaving the error unchanged is the correct way in a linear treatment of uncertainties.
- The README or documentation should include a statement of need and what sets this package apart from similar packages. This section could also be copied from the paper.
- The README or documentation should include possible directions for contribution and extension. In part, these could be taken from the "future work" section of the paper. Alternatively, the enhancements could be added as open issues to the repository.

### Further comments and suggestions - address if you like

- The tests threw errors for me with pytest (error message: `ImportError: DLL load failed while importing utils: A dynamic link library (DLL) initialization routine failed.`), but I think that was a pytest/Windows issue with Jax, so I'm just ignoring this here and trust the other reviewers on reviewing the tests.
- numpy operations do not work on scalar uncertainties (tested with `np.exp`). I don't see this as a must, but it would improve the usability of the package.
- The comparison operator `==` returns `True` for two variables with same mean but different standard deviations. This is somewhat unexpected, but I don't see a clear solution. Changing the behavior to `False` makes even less sense. Maybe throwing a warning for this case would be a good handling.
- The review list requires "Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support". While I think most of this could be handled through github PRs and issues, maybe activating discussions for this repo, providing a contact email, or some other way of getting in touch may be helpful.
