### 2025-05-06 varchasgopalaswamy/AutoUncertainties#25 @pescap
JOSS REVIEW (Reviewer #2)
This issue is related to [openjournals/joss-reviews#8037](https://github.com/openjournals/joss-reviews/issues/8037).

## Installation

- As mentioned by another Reviewer, I had trouble installing it with `pip`. The installation command is: `pip install auto-uncertainties`
- Locally, the command `pip install auto_uncertainties[docs]` did not work. It returned: `no matches found: auto_uncertainty[docs]`
- I tried installing it on my M2 MacBook Air. When importing `Uncertainty`, I get the following error:

```
File ~/miniconda3/envs/bpe/lib/python3.11/site-packages/jax/_src/lib/__init__.py:84
     80 # Before importing any C compiled modules from jaxlib, first import the CPU
     81 # feature guard module to verify that jaxlib was compiled in a way that only
     82 # uses instructions that are present on this machine.
     83 import jaxlib.cpu_feature_guard as cpu_feature_guard
---> 84 cpu_feature_guard.check_cpu_features()
     86 import jaxlib.utils as utils  # noqa: F401
     87 import jaxlib.xla_client as xla_client

RuntimeError: This version of jaxlib was built using AVX instructions, which your CPU and/or operating system do not support. This error is frequently encountered on macOS when running an x86 Python installation on ARM hardware. In this case, try installing an ARM build of Python. Otherwise, you may be able to work around this issue by building jaxlib from source.
```

I'm not sure this point is relevant, but I preferred to mention it anyway.

## Functionnality

What about support for matrices? For example, I tried:

```
A = np.random.random((10, 10))
E = np.random.random((10, 10)) * 0.1
U = Uncertainty(A, E)
```

But then:

```
np.exp(U)
```

returns:
```
Incompatible shapes for broadcasting: shapes=[(1, 100), (1, 10, 10)]
```

## CI/CD

In the `README.md`, mention the CI/CD process and how to run the tests.

## Statement of Need

Returning to the matrix case, it would be helpful to include a real or more concrete example. When I first started using the library, I naturally thought about linear system propagation errors. Including a concrete (e.g., physical) example—perhaps one that originally motivated the development of the library—would strengthen this section.

## Community Guidelines

Add contributing and community guidelines. Double-check the reviewers’ checklist to ensure that everything is complete.
