### 2025-05-13 varchasgopalaswamy/AutoUncertainties#24 @cometbeetle
Hi @JakobBD,

Thanks for taking the time to review AutoUncertainties for JOSS. I'm in the process of implementing the reviewer suggestions, and will let you know if I have any specific questions about what you wrote above.

I will mention that the `np.exp` issue you were experiencing was likely due to an issue where integer scalars cause issues with NumPy. In a PR with the JOSS-suggested changes, I have a fix where we just convert integer data types to floating point upon instantiation.
