### 2025-05-13 varchasgopalaswamy/AutoUncertainties#25 @cometbeetle
Hi @pescap,

Thank you for taking the time to review AutoUncertainties!

I'm now going through and implementing the JOSS reviewer suggestions, and will let you know if I have any specific questions about what you wrote above.
