### 2025-05-21 varchasgopalaswamy/AutoUncertainties#24 @cometbeetle
Hi again @JakobBD,

**UPDATE: The changes discussed below are now in the latest official release of AutoUncertainties!**

I wanted to let you know that I've implemented most of the changes suggested by you and the other JOSS reviewers, which you can find in [this branch](https://github.com/cometbeetle/AutoUncertainties/tree/rewrite) of my fork of AutoUncertainties. Some things may still be adjusted, however until the changes are merged into the main repository, you can install the modified version with:

```
pip install git+https://github.com/cometbeetle/AutoUncertainties.git@rewrite
```

You can also view the updated documentation at https://autouncertainties-dev1.readthedocs.io/, which syncs with the `rewrite` branch.

Below, I will summarize what I've changed with regard to your prior comments. Feel free to let me know if anything is unclear.

- **Main concern**
  * The package limitations have been added to the docs and to the README.
  * It is true that, ideally, AutoUncertainties would handle cases where dependent random variables are used, but, as mentioned in the paper, that is out of scope at this time. I will add an open GitHub issue to the repository about this in case future contributors want to try implementing something similar to what you described.

- **Other minor issues**
  * I was not able to replicate the `set_downcast_error` issue that you described, but given that many changes have been made in the `rewrite` branch, I wonder if you will no longer have that problem if you try it again.
  * I do see that the modulo operator sets the uncertainty to zero. I have left this behavior in for now, but will ask @varchasgopalaswamy if this behavior is intended, or if it should leave the uncertainty in place.
  * A proper license has been added (GNU GPLv3).
  * The statement of need has been added to the README and to the docs.

- **Further comments and suggestions**
  * I was also unfortunately unable to replicate the issues you had with the tests. I run them myself often on both Windows and macOS (ARM), and they appear run fine on GitHub Actions. If others have issues with them, I will investigate further.
  * NumPy operations (like `np.exp`) should now work on scalar uncertainty objects, as the instantiation behavior has been completely rewritten to avoid the bizarre integer / float discrepancies.
  * The `==` operator has been adjusted to raise warnings or errors (the user can decide) when equality evaluates to `True`, but the standard deviations differ.
  * I think the primary way of seeking support will be the GitHub repository, and I will ask @varchasgopalaswamy to enable the discussions feature. I believe this is now clear from the documentation, but let me know if you would like to see something more.
