### 2025-05-21 varchasgopalaswamy/AutoUncertainties#25 @cometbeetle
Hi again @pescap,

**UPDATE: The changes discussed below are now in the latest official release of AutoUncertainties!**

I wanted to let you know that I've implemented most of the changes suggested by you and the other JOSS reviewers, which you can find in [this branch](https://github.com/cometbeetle/AutoUncertainties/tree/rewrite) of my fork of AutoUncertainties. Some things may still be adjusted, however until the changes are merged into the main repository, you can install the modified version with:

```
pip install git+https://github.com/cometbeetle/AutoUncertainties.git@rewrite
```

You can also view the updated documentation at https://autouncertainties-dev1.readthedocs.io/, which syncs with the `rewrite` branch.

Below, I will summarize what I've changed with regard to your prior comments. Feel free to let me know if anything is unclear.

- **Installation**
  * I was not able to replicate your installation issue on my M1 Pro MacBook Pro, but I wonder if it has something to do with your Anaconda install. I would suggest trying the installation again with something like `virtualenv` or [uv](https://docs.astral.sh/uv/) to manage the virtual environment. Also, check to see if you are using an x86 Python installation, as according to your error message, that could be a source of issues too. If nothing works, let me know, and I'll try to investigate further.

- **Functionality**
  * Matrices should be supported now. I'm not sure why it was not working before, but I was able to run the example you provided without issues using the `rewrite` branch.

- **CI/CD**
  * The CI/CD process is now mentioned in both the documentation and the README.

- **Statement of Need**
  * I have not modified the statement of need that much, but I will ping @varchasgopalaswamy about a potential concrete example that motivated the development of AutoUncertainties.

- **Community Guidelines**
  * The docs now include a short section about contribution. Let me know if you think it should be expanded.
