### 2025-05-21 varchasgopalaswamy/AutoUncertainties#21 @cometbeetle
Hi @damar-wicaksono,

**UPDATE: The changes discussed below are now in the latest official release of AutoUncertainties!**

I wanted to let you know that I've implemented most of the changes suggested by you and the other JOSS reviewers, which you can find in [this branch](https://github.com/cometbeetle/AutoUncertainties/tree/rewrite) of my fork of AutoUncertainties. Some things may still be adjusted, however until the changes are merged into the main repository, you can install the modified version with:

```
pip install git+https://github.com/cometbeetle/AutoUncertainties.git@rewrite
```

You can also view the updated documentation at https://autouncertainties-dev1.readthedocs.io/, which syncs with the `rewrite` branch.

Below, I will summarize what I've changed with regard to your prior comments. I'll also add a separate comment to your other issue (#22) about the functionality. Feel free to let me know if anything is unclear.

- **Statement of Need**
  * Fixed the typo.

- **Prior Work**
  * Made the limitations of `uncertainties` more explicit.
  * Reworded the portion talking about "transparent" behavior. It should now be easier to understand.
  * Simplified the `uncertainties` code example, and indicated which version was used.

- **Implementation**
  * The uncertainty propagation formula has been clarified, and now includes the proper summation.
  * The class structure of `Uncertainty` objects has been simplified. Specifically, `VectorUncertainty` and `ScalarUncertainty` have been **removed** in favor of a more flexible `Uncertainty` class. This is a major change, but it should now be easier to understand the instantiation logic, as well as the method implementations.
  * I will also mention that the difference between a vector `Uncertainty` object and an array of scalar `Uncertainty` objects has to do mainly with computational efficiency and proper management of array calculation. It is much faster to perform NumPy operations on arrays of numbers (since this can be often done in parallel) than it is to individual operate on object elements of an array. Hopefully this is more clear with the simplified class structure, but let me know if that isn't the case.
  * Added an example of AutoUncertainties using standard NumPy functions automatically, as suggested.
  * Moved information about `nominal_values` and `std_devs` to the documentation. These are mainly to support certain uses of AutoUncertainties in other software.
  * Clarified that "error" means the same thing as "standard deviation" for this paper.
  * Removed rounding display example, and put all relevant information in the docs.
  * Removed `set_downcast_error` example, and put all relevant information in the docs.
  * The `from_sequence` method could probably be moved inside `__new__` or marked private, but is kept to maintain compatibility. Its primary purpose is to intercept and handle sequences containing `Quantity` objects during instantiation.
  * `from_quantities` could probably be moved inside `__new__` or marked private, but for now it is kept for compatibility.
  * Fixed typo in the Pandas section.

- **Current Limitations and Future Work**
  * The i.i.d. assumption is now mentioned in the Statement of Need.
  * Modified the section to mention that the dependence assumption affects most binary operations performed on the same `Uncertainty` object.
  * Emphasized that the assumption leads to common pitfalls, since users are not prevented from inconsistently manipulating objects.
  * Fixed typo.
  * Moved the examples and most of the discussion of the unexpected behavior to the documentation, and slightly expanded the workaround information. The documentation now mentions the subtleties of propagating uncertainty for variables that are not *fully* correlated.
  * Removed the section about typing (as this is now mostly resolved in the rewrite).
  * In case I missed something, let me know if the re-worked limitations section needs any adjustments or additions!
