### 2025-05-21 varchasgopalaswamy/AutoUncertainties#22 @cometbeetle
Hi again @damar-wicaksono,

**UPDATE: The changes discussed below are now in the latest official release of AutoUncertainties!**

This is the second part of my comment from #21. Just for easy reference, I'll mention again that you can install the modified version of AutoUncertainties with:

```
pip install git+https://github.com/cometbeetle/AutoUncertainties.git@rewrite
```

And the updated docs are at: https://autouncertainties-dev1.readthedocs.io/, which syncs with the `rewrite` branch.

Below, I will summarize what I've changed with regard to your prior comments. Feel free to let me know if anything is unclear.

- **Installation**
  * The correct package is `auto-uncertainties` on PyPi, so the documentation has been adjusted to use that name instead of the underscored version (though, I believe both names point to the correct package).

- **Functionality**
  * The relative uncertainty is now always `nan` when the central value is zero.
  * The documentation for `set_display_rounding` has been clarified to correctly describe the implementation.
  * Calling `np.array()` on a scalar `Uncertainty` object will now raise a `DowncastWarning`, and cast down to a 0-D NumPy array. This makes the functionality more consistent, and prevents creating NumPy arrays of `Uncertainty` objects.
  * The `Uncertainty` constructor now accepts `Sequence` objects for both the value and the error, and implements better input validation and type checking.
  * The issue where a NaN relative uncertainty causes problems with certain multiplication and division operations has now been fixed by expanding the formula to cancel out certain problematic terms.
