### 2025-06-15 #8037 @cometbeetle
Hi @osorensen,

Thanks for reaching out. I have responded to all the reviewers' comments in their respective pull requests ([1](https://github.com/varchasgopalaswamy/AutoUncertainties/issues/21), [2](https://github.com/varchasgopalaswamy/AutoUncertainties/issues/22), [3](https://github.com/varchasgopalaswamy/AutoUncertainties/issues/24), [4](https://github.com/varchasgopalaswamy/AutoUncertainties/issues/25)), and have implemented (almost all) the necessary changes in the [rewrite branch](https://github.com/cometbeetle/AutoUncertainties/tree/rewrite) of my fork of AutoUncertainties. The original repository owner (@varchasgopalaswamy) has been out of the country since the beginning of May, so my changes have not yet been merged into the main branch. I expect him to return soon, though.

At the moment, I'm waiting on responses from the reviewers. There are still a small number of adjustments that I need to ask @varchasgopalaswamy about; however, the vast majority of the reviewers' suggestions have been implemented.

If there's anything I need to do in the meantime, please let me know. Thanks!
