### 2025-06-15 #8037 @cometbeetle
Also, if it's helpful, the latest PDF generations of the paper can be found in the [GitHub Actions artifacts](https://github.com/cometbeetle/AutoUncertainties/actions/workflows/draft-pdf.yml?query=branch%3Arewrite) under my fork.

Once my changes are merged into the main repository, the `generate pdf` command for the editorial bot should work fine. I'll be sure to update this issue once that happens.
