### 2025-06-17 #8037 @cometbeetle
@osorensen Just wanted to let you know that my updates have been merged into the AutoUncertainties stable branch, and the latest release now includes everything.
