### 2025-06-18 varchasgopalaswamy/AutoUncertainties#24 @JakobBD
Thanks for the response.
All the raised points seem adequately addressed, except for the two minor ones you wanted to discuss with @varchasgopalaswamy .
It would be great if you could still do that.
As those are minor points, I am closing the issue already.
