### 2025-06-18 varchasgopalaswamy/AutoUncertainties#24 @cometbeetle
Hi @JakobBD,

Thanks for getting back to us! About the two minor points, I just wanted to say that:

1.  @varchasgopalaswamy has said to me that he will enable the discussions feature.
2. We did discuss the modulo operator, and are considering whether to leave its current behavior or have it keep the uncertainty unchanged. In a future release, its behavior will be clarified in the documentation according to what we decide.
