### 2025-06-19 varchasgopalaswamy/AutoUncertainties#21 @damar-wicaksono
@cometbeetle,

Thank you very much for going through my comments and addressing them! I can confirm that everything is in order.

I noticed a few minor points that might be worth addressing in the latest draft:

- ll. 13-14 independent and identically distributed (i.i.d): I don't think identically distributed is the underlying assumption of the package. Each random variable may have different mean (central value) and error (standard deviation); they are by construction not identical.
- l. 39 "...differentiaion...": typo (missing 't')
- ll. 68, 71 "...where the programmer assumes..." and "...will not prevent users from...": For consistency, you might consider using one term throughout (either "programmer" or "user").
