### 2025-06-19 varchasgopalaswamy/AutoUncertainties#21 @cometbeetle
@damar-wicaksono Thanks for catching those last few things! I'll make a PR with the appropriate changes, and discuss with @varchasgopalaswamy. Everything should be complete by the end of the week. I'll let you know when the paper is finalized in the next minor release.
