### 2025-06-26 5cafbdd
Trying to fix compilation error
https://github.com/openjournals/joss-reviews/issues/8037
