### 2025-06-30 ff57099
Trying to fix compilation error (#30)
https://github.com/openjournals/joss-reviews/issues/8037
