### 2025-06-30 ba7df0d
Trying to fix compilation error (#30) (#31)
https://github.com/openjournals/joss-reviews/issues/8037

Co-authored-by: Øystein Sørensen <oystein_sorensen@hotmail.com>
