### 2025-07-01 #8037 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
