### 2025-07-03 ceac174
add CITATION.cff
